requirements.txt
amqp==2.6.1
asgiref==3.4.1
beat==0.3.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==4.4.7
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-appconf==1.0.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-classy-tags==2.0.0
django-filer==2.0.2
django-formtools==2.3
django-imagekit==4.0.2
django-import-export==2.5.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-mail-template==0.1.11
django-mptt==0.13.3
django-polymorphic==3.0.0
django-tagging-autocomplete==0.5.1
django-taggit==1.5.1
django-taggit-autocomplete-modified==0.1.0b6
django-taggit-autosuggest==0.3.8
django-timezone-field==4.2.1
django-tinymce==3.3.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-page-meta==1.0.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.2
Jinja2==3.0.1
kombu==4.6.11
lxml==4.6.3
pilkit==2.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
pyseoanalyzer==4.0.3
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
razorpay==1.2.0
redis==3.5.3
requests==2.26.0
sentry-sdk==1.3.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
structlog==21.1.0
tablib==3.0.0
Unidecode==1.1.2
urllib3==1.26.6
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
django-admin-honeypot==1.1.0
qrcode==7.3.1
django-otp==1.1.1